refractor: Move the list
Moved the common list of valid city, month and day names outside the get_filters function.
This is to avoid code duplication.
I defined them as global constants at the beginning of the script.
This change helps maintain consistency and makes it easier to update the valid names in the future.